Refactor MySQL support to better match what Omnibus supports
