+++ manifests/params.pp
@@ -136,12 +136,17 @@ class gitlab::params {
   #
   # 5. Omnibus customization
   # ==========================
-  $postgresql_enable  = undef # (default: true)
+  $postgresql_enable  = true
   $mysql_enable       = false
-  $mysql_host         = '127.0.0.1'
-  $mysql_port         = '3306'
-  $mysql_username     = 'git'
-  $mysql_password     = undef
+  $db_adapter         = undef # (default: 'postgresql')
+  $db_encoding        = undef # (default: 'unicode')
+  $db_database        = undef # (default: 'gitlabhq_production')
+  $db_pool            = undef # (default: 10)
+  $db_username        = undef # (default: 'gitlab')
+  $db_password        = undef # (default: nil)
+  $db_host            = undef # (default: nil)
+  $db_port            = undef # (default: 5432)
+  $db_socket          = undef # (default: nil)
 
   $redis_port       = undef # (default: 6379)
   $postgresql_port  = undef # (default: 5432)
