+++ templates/gitlab-puppet.rb.erb
@@ -94,16 +94,20 @@ gitlab_rails['ldap_admin_group'] = '<%= @ldap_admin_group %>'
 #
 # 5. Omnibus customization
 # ==========================
-gitlab_rails['postgres'] = <%= @postgresql_enable %>
-<% if @mysql_enable -%>
-gitlab_rails['db_adapter'] = 'mysql2'
-gitlab_rails['db_encoding'] = 'utf8'
-gitlab_rails['db_host'] = '<%= @mysql_host %>'
-gitlab_rails['db_port'] = '<%= @mysql_port %>'
-gitlab_rails['db_username'] = '<%= @mysql_username %>'
-gitlab_rails['db_password'] = '<%= @mysql_password %>'
-<% end -%><% if @redis_port      -%>redis['port'] = <%= @redis_port %>
-<% end -%><% if @postgresql_port -%>postgresql['port'] = <%= @postgresql_port %>
+postgresql['enable'] = <%= @postgresql_enable %>
+<% if @postgresql_enable == false || @mysql_enable == true -%>
+<% if @db_adapter -%>gitlab_rails['db_adapter'] = '<%= @db_adapter %>'
+<% end %><% if @db_encoding %>gitlab_rails['db_encoding'] = '<%= @db_encoding %>'
+<% end %><% if @db_database -%>gitlab_rails['db_database'] = '<%= @db_database %>'
+<% end %><% if @db_pool -%>gitlab_rails['db_pool'] = <%= @db_pool %>
+<% end %><% if @db_username -%>gitlab_rails['db_username'] = '<%= @db_username %>'
+<% end %><% if @db_password -%>gitlab_rails['db_password'] = '<%= @db_password %>'
+<% end %><% if @db_host -%>gitlab_rails['db_host'] = '<%= @db_host %>'
+<% end %><% if @db_socket -%>gitlab_rails['db_socket'] = <%= @mysql_socket %>
+<% end -%>
+<% end %><% if @db_port -%>gitlab_rails['db_port'] = '<%= @db_port %>'
+<% end %><% if @redis_port      -%>redis['port'] = <%= @redis_port %>
+<% end %><% if @postgresql_port -%>postgresql['port'] = <%= @postgresql_port %>
 <% end %><% if @unicorn_port     -%>unicorn['port'] = <%= @unicorn_port %>
 <% end %><% if @git_data_dir     -%>git_data_dir = '<%= @git_data_dir  %>'
 <% end %><% if @gitlab_username  -%>user['username'] = '<%= @gitlab_username  %>'
